chore: Render dashbord widgets asynchronously
