+++ app/config/packages/doctrine.yaml
@@ -27,6 +27,19 @@ doctrine:
                 DATE: App\Doctrine\Extensions\DateFunction
                 FROM_UNIXTIME: App\Doctrine\Extensions\FromUnixTimeFunction
 
+when@dev:
+    doctrine:
+        orm:
+            result_cache_driver:
+                type: pool
+                pool: doctrine.result_cache_pool
+
+    framework:
+        cache:
+            pools:
+                doctrine.result_cache_pool:
+                    adapter: cache.app
+
 when@prod:
     doctrine:
         orm:
