+++ app/templates/dashboard/stats-by-in-out.html.twig
@@ -1,0 +1,135 @@
+<turbo-frame id="dashboard-stats-by-in-out">
+        <div class="card-body h-100">
+            <div class="row">
+                <div class="col-md-2 d-flex align-items-center">
+                    <form id="domainForm">
+                        <div class="form-group">
+                            <label for="domainSelect">{{'Entities.Domain.Name' | trans }}
+                                :</label>
+                            <select class="form-control" id="domainSelect" name="domain">
+                                <option value="All">{{'Generics.labels.all' | trans }}</option>
+                                {% for domain in domains %}
+                                    <option value="{{ domain.id }}">{{ domain.domain }}</option>
+                                {% endfor %}
+                            </select>
+                        </div>
+                    </form>
+                </div>
+                <div class="col-md-5" style="height: 250px;">
+                    <canvas id="msgsChart"></canvas>
+                </div>
+                <div class="col-md-5" style="height: 250px;">
+                    <canvas id="outMsgsChart"></canvas>
+                </div>
+            </div>
+            <a href="#" data-toggle="modal" data-target="#statsTableModal" style="position: absolute; top: 10px; right: 10px;">
+                <i class="fas fa-expand-arrows-alt fa-2x"></i>
+            </a>
+        </div>
+
+    {% if is_granted('ROLE_ADMIN') %}
+        <div class="modal fade" id="statsTableModal" tabindex="-1" role="dialog" aria-labelledby="statsTableModalLabel" aria-hidden="true">
+            <div class="modal-dialog modal-xl" role="document">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title" id="statsTableModalLabel">{{ 'Generics.labels.statTable'|trans }}</h5>
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                            <span aria-hidden="true">&times;</span>
+                        </button>
+                    </div>
+                    <div class="modal-body">
+                        {% include '/dashboard/_stats-in-out_table.html.twig' %}
+                    </div>
+                </div>
+            </div>
+        </div>
+    {% endif %}
+
+<script>
+document.addEventListener("turbo:frame-load", function () {
+    var ctxMsgs = document.getElementById('msgsChart').getContext('2d');
+    var msgsData = {
+        labels: [
+            '{{ 'Generics.labels.unblockedMessages'|trans }}', '{{ 'Generics.labels.blockedMessages'|trans }}'
+        ],
+        datasets: [
+            {
+                data: [
+                    {{ data['All'].totalMsgCount - data['All'].totalMsgBlockedCount }},{{ data['All'].totalMsgBlockedCount }}
+                ],
+                backgroundColor: [
+                    '#36A2EB', // Blue
+                    '#FF6384' // Red
+                ]
+            }
+        ]
+    };
+
+    var msgsChart = new Chart(ctxMsgs, {
+        type: 'doughnut',
+        data: msgsData,
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            legend: {
+                position: 'right'
+            },
+            title: {
+                display: true,
+                text: '{{ 'Generics.labels.incomingMails'|trans }}'
+            }
+        }
+    });
+
+    var ctxOutMsgs = document.getElementById('outMsgsChart').getContext('2d');
+    var outMsgsData = {
+        labels: [
+            '{{ 'Generics.labels.unblockedMessages'|trans }}', '{{ 'Generics.labels.blockedMessages'|trans }}'
+        ],
+        datasets: [
+            {
+                data: [
+                {{ data['All'].totalOutMsgCount - data['All'].totalOutMsgBlockedCount }},{{ data['All'].totalOutMsgBlockedCount }}
+                ],
+                backgroundColor: [
+                    '#36A2EB', // Blue
+                    '#FF6384' // Red
+                ]
+            }
+        ]
+    };
+
+    var outMsgsChart = new Chart(ctxOutMsgs, {
+        type: 'doughnut',
+        data: outMsgsData,
+        options: {
+            responsive: true,
+            maintainAspectRatio: false,
+            legend: {
+                                                        position: 'right'
+            },
+            title: {
+                display: true,
+                text: '{{ 'Generics.labels.outgoingMails'|trans }}'
+            }
+        }
+    });
+
+    var data = {{ data|json_encode|raw }};
+    document.getElementById('domainSelect').addEventListener('change', function () {
+        var selectedDomain = this.value;
+        var domainData = data[selectedDomain];
+
+    // Update the charts with the new data
+        msgsChart.data.datasets[0].data = [
+            domainData.totalMsgCount - domainData.totalMsgBlockedCount,
+            domainData.totalMsgBlockedCount
+        ];
+        msgsChart.update();
+
+        outMsgsChart.data.datasets[0].data = [domainData.totalOutMsgCount, domainData.totalOutMsgBlockedCount];
+        outMsgsChart.update();
+    });
+});
+</script>
+</turbo-frame>
